backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
tzdata>=2024.2
motor==3.7.1
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
emergentintegrations
Pillow
reportlab>=4.0.0
openpyxl>=1.1.0
python-docx>=1.1.0
fal-client>=0.7.0
aiohttp
woocommerce
httpx
tenacity
apscheduler>=3.10.4
psutil>=5.9.0

# Amazon SP-API dependencies
python-amazon-sp-api>=1.2.0

# Web Scraping Libraries
pytrends>=4.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
aiohttp-retry
fake-useragent>=1.4.0
aiohttp-socks>=0.8.0

# Testing & Quality Tools
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
freezegun>=1.2.0
respx>=0.20.0
coverage>=7.0.0
faker>=24.0.0

# Code Quality
ruff>=0.2.0
mypy>=1.8.0
black>=24.1.1
isort>=5.13.2

# Amazon SP-API Integration
boto3>=1.34.0
cryptography>=41.0.7
httpx>=0.25.2

# Development
# pre-commit>=3.6.0   # Removed - development only
pyjwt>=2.8.0
bcrypt>=4.1.0
